fixed help

+++ src/dp/traceback.h
@@ -75,7 +75,7 @@ int have_diag(const _matrix &dp,
 
 		if (dp(i, j) == match_score + dp(i - 1, j - 1)) {
 			if (log)
-				printf("i=%i j=%i score=%i subject=%c query=%c\n", i, j, dp(i, j), value_traits.alphabet[subject[j - 1]], value_traits.alphabet[query[i - 1]]);
+				printf("i=%i j=%i score=%i subject=%c query=%c\n", i, j, dp(i, j), value_traits.alphabet[(int)subject[j - 1]], value_traits.alphabet[(int)query[i - 1]]);
 			++l;
 			--i;
 			--j;
